Move likelihood dist to distributions submodule
